refactor: split integration test into each command
